+++ Tests/SuperwallKitTests/Config/ConfigManagerTests.swift
@@ -13,7 +13,10 @@ import XCTest
 final class ConfigManagerTests: XCTestCase {
   func test_refreshConfiguration() async {
     let dependencyContainer = DependencyContainer()
-    let network = NetworkMock(factory: dependencyContainer)
+    let network = NetworkMock(
+      options: SuperwallOptions(),
+      factory: dependencyContainer
+    )
     let newConfig: Config = .stub()
       .setting(\.buildId, to: "123")
     network.configReturnValue = .success(newConfig)
@@ -42,16 +45,17 @@ final class ConfigManagerTests: XCTestCase {
     let dependencyContainer = DependencyContainer()
 
     let paywall = Paywall.stub()
-      .setting(\.productItems, to: [.init(name: "abc", type: .appStore(.init(store: .appStore, id: "abc")))])
+      .setting(\.products, to: [.init(name: "abc", type: .appStore(.init(store: .appStore, id: "abc")), entitlements: [])])
     let config: Config = Config(
       buildId: "buildId",
       triggers: [
         .init(
-          eventName: "event",
-          rules: [.stub()]
+          placementName: "event",
+          audiences: [.stub()]
         )
       ],
       paywalls: [paywall],
+      entitlements: [],
       logLevel: 2,
       locales: ["fr"],
       appSessionTimeout: 2202,
@@ -73,7 +77,10 @@ final class ConfigManagerTests: XCTestCase {
       variant: variant
     )
     let dependencyContainer = DependencyContainer()
-    let network = NetworkMock(factory: dependencyContainer)
+    let network = NetworkMock(
+      options: SuperwallOptions(),
+      factory: dependencyContainer
+    )
     let storage = StorageMock()
     let configManager = ConfigManager(
       options: SuperwallOptions(),
@@ -100,7 +107,10 @@ final class ConfigManagerTests: XCTestCase {
 
   func test_loadAssignments_noConfig() async {
     let dependencyContainer = DependencyContainer()
-    let network = NetworkMock(factory: dependencyContainer)
+    let network = NetworkMock(
+      options: SuperwallOptions(),
+      factory: dependencyContainer
+    )
     let storage = StorageMock()
     let configManager = ConfigManager(
       options: SuperwallOptions(),
@@ -128,7 +138,10 @@ final class ConfigManagerTests: XCTestCase {
 
   func test_loadAssignments_noTriggers() async {
     let dependencyContainer = DependencyContainer()
-    let network = NetworkMock(factory: dependencyContainer)
+    let network = NetworkMock(
+      options: SuperwallOptions(),
+      factory: dependencyContainer
+    )
     let storage = StorageMock()
     let configManager = ConfigManager(
       options: SuperwallOptions(),
@@ -151,7 +164,10 @@ final class ConfigManagerTests: XCTestCase {
 
   func test_loadAssignments_saveAssignmentsFromServer() async {
     let dependencyContainer = DependencyContainer()
-    let network = NetworkMock(factory: dependencyContainer)
+    let network = NetworkMock(
+      options: SuperwallOptions(),
+      factory: dependencyContainer
+    )
     let storage = StorageMock()
     let configManager = ConfigManager(
       options: SuperwallOptions(),